requirements.txt
beautifulsoup4==4.12.2
EbookLib==0.17.1
google_api_python_client==2.88.0
google_auth_oauthlib==0.8.0
langchain==0.0.194
numpy==1.23.5
openai==0.27.6
openai_whisper==20230314
protobuf==4.23.2
PyPDF2==3.0.1
pytube==12.1.2
requests==2.29.0
spotipy==2.23.0
whisper==1.1.10
youtube_transcript_api==0.5.0
yt_dlp==2023.3.4
